data file headers for each sensor

+++ projects/f207/Core/Inc/utils.h
@@ -11,6 +11,7 @@
 #include "fatfs.h"
 #include "cJSON.h"
 #include "usart.h"
+#include "spi.h"
 
 
 int parse_config(const char *config_str, driver *driver_list,sensor *sensor_list,monitor *monitor_list,
@@ -25,7 +26,7 @@ int append_file(FIL *target_file, char *data, UINT btw);
 int mount_sd(FATFS* FatFs);
 int close_file(FIL *target_file);
 float get_sensorval(sensor *current_sensor);
-uint16_t get_mcp3208_adcval(int channel, int cs, SPI_HandleTypeDef *spiHandle);
+uint16_t get_mcp3208_adcval(int channel, uint16_t cs, SPI_HandleTypeDef *spiHandle);
 void filter_and_decimate(float *sensor_vals, int sensor_count);
 
 #endif /* INC_UTILS_H_ */
